tagsが膨れ過ぎたので、視覚的な操作はvisualsに分離 #297
