+++ app/javascript/views/reports_index.vue
@@ -50,7 +50,7 @@
 import ImageShow from '../components/image_show.vue'
 import LocatorsIndex from '../components/locators_index.vue'
 import params from '../modules/params'
-import tags from '../modules/tags'
+import visuals from '../modules/visuals'
 
 export default {
   name: 'RoomsShow',
@@ -95,7 +95,7 @@ export default {
     this.printedAt = params.formatDate()
   },
   mounted() {
-    tags.hideBars()
+    visuals.hideBars()
   }
 }
 </script>
